+++ dashboard/package.json
@@ -20,16 +20,14 @@
     "axios": "^1.6.2",
     "axios-mock-adapter": "^1.22.0",
     "chance": "1.1.11",
+    "d3": "^7.9.0",
     "date-fns": "2.30.0",
-    "graphology": "^0.26.0",
-    "graphology-layout-force": "^0.2.4",
     "highlight.js": "^11.11.1",
     "js-md5": "^0.8.3",
     "lodash": "4.17.21",
     "marked": "^15.0.7",
     "pinia": "2.1.6",
     "remixicon": "3.5.0",
-    "sigma": "^3.0.1",
     "vee-validate": "4.11.3",
     "vite-plugin-vuetify": "1.0.2",
     "vue": "3.3.4",
